enhancement: Also something something about client navigation bobert kms przez cb kiedys <333
